requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
oauthlib==3.1.1
Pillow==8.3.1
PySocks==1.7.1
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.6
